Refactored and reworked Danuar Reliquary

+++ game-server/data/scripts/system/handlers/ai/instance/danuarReliquary/MaliciousIceStormAI.java
@@ -3,10 +3,12 @@ package ai.instance.danuarReliquary;
 import com.aionemu.gameserver.ai.AIActions;
 import com.aionemu.gameserver.ai.AIName;
 import com.aionemu.gameserver.ai.NpcAI;
+import com.aionemu.gameserver.model.skill.NpcSkillEntry;
 import com.aionemu.gameserver.utils.ThreadPoolManager;
 
 /**
  * @author Ritsu
+ * @modified Estrayl October 29th, 2017.
  */
 @AIName("malicious_ice_storm")
 public class MaliciousIceStormAI extends NpcAI {
@@ -14,27 +16,12 @@ public class MaliciousIceStormAI extends NpcAI {
 	@Override
 	protected void handleSpawned() {
 		super.handleSpawned();
-		iceStorm();
+		ThreadPoolManager.getInstance().schedule(() -> AIActions.useSkill(MaliciousIceStormAI.this, 21180), 1000);
 	}
 
-	private void iceStorm() {
-		ThreadPoolManager.getInstance().schedule(new Runnable() {
-
-			@Override
-			public void run() {
-				AIActions.useSkill(MaliciousIceStormAI.this, 21180);
-				despawn();
-			}
-		}, 100);
-	}
-
-	private void despawn() {
-		ThreadPoolManager.getInstance().schedule(new Runnable() {
-
-			@Override
-			public void run() {
-				getOwner().getController().delete();
-			}
-		}, 2100);
+	@Override
+	public void onEndUseSkill(NpcSkillEntry usedSkill) {
+		if (usedSkill.getSkillId() == 21180)
+			getOwner().getController().delete();
 	}
 }
