+++ game-server/data/scripts/system/handlers/ai/instance/danuarReliquary/VengefulOrbAI.java
@@ -1,40 +1,38 @@
 package ai.instance.danuarReliquary;
 
+import com.aionemu.commons.utils.Rnd;
 import com.aionemu.gameserver.ai.AIActions;
 import com.aionemu.gameserver.ai.AIName;
 import com.aionemu.gameserver.ai.NpcAI;
-import com.aionemu.gameserver.utils.ThreadPoolManager;
+import com.aionemu.gameserver.model.skill.NpcSkillEntry;
+import com.aionemu.gameserver.skillengine.model.Effect;
 
 /**
  * @author Ritsu
+ * @modified Estrayl October 28th, 2017.
  */
 @AIName("vengeful_orb")
 public class VengefulOrbAI extends NpcAI {
 
+	/**
+	 * Currently capped the damage value. Seems to be some defensive calculations are not even retail.
+	 * Damage on retail: 12000 - 13000
+	 * Damage here: 17000+
+	 */
+	@Override
+	public int modifyOwnerDamage(int damage, Effect effect) {
+		return damage > 12500 ? Rnd.get(12000, 13000) : damage;
+	}
+
 	@Override
 	protected void handleSpawned() {
 		super.handleSpawned();
-		vengefulOrb();
+		AIActions.useSkill(VengefulOrbAI.this, 21178);
 	}
 
-	private void vengefulOrb() {
-		ThreadPoolManager.getInstance().schedule(new Runnable() {
-
-			@Override
-			public void run() {
-				AIActions.useSkill(VengefulOrbAI.this, 21178);
-				despawn();
-			}
-		}, 100);
-	}
-
-	private void despawn() {
-		ThreadPoolManager.getInstance().schedule(new Runnable() {
-
-			@Override
-			public void run() {
-				getOwner().getController().delete();
-			}
-		}, 13000);
+	@Override
+	public void onEndUseSkill(NpcSkillEntry usedSkill) {
+		if (usedSkill.getSkillId() == 21178)
+			getOwner().getController().delete();
 	}
 }
